Move range key graphs in seperate package
