+++ src/main/java/de/hpi/msd/salsa/EdgeToAdjacencyApp.java
@@ -1,6 +1,8 @@
 package de.hpi.msd.salsa;
 
 import de.hpi.msd.salsa.graph.*;
+import de.hpi.msd.salsa.graph.rangeKey.RangeKeyGraph;
+import de.hpi.msd.salsa.graph.rangeKey.SampleKeyValueGraph;
 import de.hpi.msd.salsa.processor.EdgeProcessor;
 import de.hpi.msd.salsa.processor.RangeKeyProcessor;
 import de.hpi.msd.salsa.processor.SamplingEdgeProcessor;
